+++ tests/allocs.rs
@@ -14,6 +14,7 @@ use std::cell::Cell;
 use std::hint::black_box;
 use std::thread_local;
 
+use blazesym::normalize::NormalizeOpts;
 use blazesym::normalize::Normalizer;
 use blazesym::Addr;
 
@@ -66,12 +67,20 @@ fn normalize_process() {
             libc::dlopen as Addr,
             libc::fopen as Addr,
             normalize_process as Addr,
-            Normalizer::normalize_user_addrs_sorted as Addr,
+            Normalizer::normalize_user_addrs as Addr,
         ];
         let () = addrs.sort();
 
+        let opts = NormalizeOpts {
+            sorted_addrs: true,
+            ..Default::default()
+        };
         let normalized = normalizer
-            .normalize_user_addrs_sorted(black_box(0.into()), black_box(addrs.as_slice()))
+            .normalize_user_addrs_opts(
+                black_box(0.into()),
+                black_box(addrs.as_slice()),
+                black_box(&opts),
+            )
             .unwrap();
         assert_eq!(normalized.meta.len(), 2);
         assert_eq!(normalized.outputs.len(), 5);